refactor(schedules): relocate schedule details view to schedule-details-page
Move the schedule route shell out of domain-schedules into a top-level view,
rename schedule-detail-* to schedule-details-* for consistency, and place the
tab allowlist config under config/schedule-details-tabs.config.ts.
