--- src/views/domain-schedules/schedule-detail-page-tab-content/schedule-detail-page-tab-content.types.ts
@@ -1,5 +1,0 @@
-import { type ScheduleDetailPageTabsParams } from '@/views/domain-schedules/schedule-detail-page-tabs/schedule-detail-page-tabs.types';
-
-export type Props = {
-  params: ScheduleDetailPageTabsParams;
-};
