--- src/views/domain-schedules/schedule-detail-page/schedule-detail-page.tsx
@@ -1,5 +1,0 @@
-import { type Props } from './schedule-detail-page.types';
-
-export default function ScheduleDetailPage({ children }: Props) {
-  return <>{children}</>;
-}
