+++ src/views/schedule-details-page/schedule-details-page-tab-content/__tests__/schedule-details-page-tab-content.test.tsx
@@ -2,8 +2,8 @@ import React from 'react';
 
 import { render, screen } from '@/test-utils/rtl';
 
-import { type ScheduleDetailPageTabsParams } from '../../schedule-detail-page-tabs/schedule-detail-page-tabs.types';
-import ScheduleDetailPageTabContent from '../schedule-detail-page-tab-content';
+import { type ScheduleDetailsPageTabsParams } from '../../schedule-details-page-tabs/schedule-details-page-tabs.types';
+import ScheduleDetailsPageTabContent from '../schedule-details-page-tab-content';
 
 jest.mock('next/navigation', () => ({
   ...jest.requireActual('next/navigation'),
@@ -12,7 +12,7 @@ jest.mock('next/navigation', () => ({
   }),
 }));
 
-describe(ScheduleDetailPageTabContent.name, () => {
+describe(ScheduleDetailsPageTabContent.name, () => {
   it('renders placeholder for details tab', () => {
     setup({ scheduleTab: 'details' });
     expect(screen.getByText(/Details/i)).toBeInTheDocument();
@@ -28,7 +28,7 @@ describe(ScheduleDetailPageTabContent.name, () => {
     expect(() =>
       setup({
         scheduleTab:
-          'unknown-tab' as ScheduleDetailPageTabsParams['scheduleTab'],
+          'unknown-tab' as ScheduleDetailsPageTabsParams['scheduleTab'],
       })
     ).toThrow('NEXT_NOT_FOUND');
     expect(notFound).toHaveBeenCalled();
@@ -38,10 +38,10 @@ describe(ScheduleDetailPageTabContent.name, () => {
 function setup({
   scheduleTab = 'details',
 }: {
-  scheduleTab?: ScheduleDetailPageTabsParams['scheduleTab'];
+  scheduleTab?: ScheduleDetailsPageTabsParams['scheduleTab'];
 } = {}) {
   render(
-    <ScheduleDetailPageTabContent
+    <ScheduleDetailsPageTabContent
       params={{
         domain: 'test-domain',
         cluster: 'test-cluster',
