+++ src/views/schedule-details-page/schedule-details-page-tab-content/schedule-details-page-tab-content.styles.ts
