+++ src/views/schedule-details-page/schedule-details-page-tab-content/schedule-details-page-tab-content.types.ts
@@ -1,0 +1,5 @@
+import { type ScheduleDetailsPageTabsParams } from '@/views/schedule-details-page/schedule-details-page-tabs/schedule-details-page-tabs.types';
+
+export type Props = {
+  params: ScheduleDetailsPageTabsParams;
+};
