+++ src/views/schedule-details-page/schedule-details-page.tsx
@@ -1,0 +1,5 @@
+import { type Props } from './schedule-details-page.types';
+
+export default function ScheduleDetailsPage({ children }: Props) {
+  return <>{children}</>;
+}
