+++ src/views/schedule-details-page/schedule-details-page.types.ts
@@ -1,12 +1,12 @@
 import type React from 'react';
 
-export type ScheduleDetailPageLayoutParams = {
+export type ScheduleDetailsPageLayoutParams = {
   domain: string;
   cluster: string;
   scheduleId: string;
 };
 
 export type Props = {
-  params: ScheduleDetailPageLayoutParams;
+  params: ScheduleDetailsPageLayoutParams;
   children: React.ReactNode;
 };
